refactor: extract filter dropdown
